refactor(api): move learning-content-repository into src/shared
