+++ api/src/shared/infrastructure/repositories/learning-content-repository.js
@@ -1,17 +1,17 @@
 import _ from 'lodash';
 
-import { knex } from '../../../db/knex-database-connection.js';
-import * as campaignRepository from '../../../src/prescription/campaign/infrastructure/repositories/campaign-repository.js';
-import { NoSkillsInCampaignError, NotFoundError } from '../../../src/shared/domain/errors.js';
-import { CampaignLearningContent } from '../../../src/shared/domain/models/CampaignLearningContent.js';
-import { LearningContent } from '../../../src/shared/domain/models/LearningContent.js';
-import * as areaRepository from '../../../src/shared/infrastructure/repositories/area-repository.js';
-import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
-import * as frameworkRepository from '../../../src/shared/infrastructure/repositories/framework-repository.js';
-import * as skillRepository from '../../../src/shared/infrastructure/repositories/skill-repository.js';
-import * as thematicRepository from '../../../src/shared/infrastructure/repositories/thematic-repository.js';
-import * as tubeRepository from '../../../src/shared/infrastructure/repositories/tube-repository.js';
-import * as learningContentConversionService from '../../domain/services/learning-content/learning-content-conversion-service.js';
+import { knex } from '../../../../db/knex-database-connection.js';
+import * as learningContentConversionService from '../../../../lib/domain/services/learning-content/learning-content-conversion-service.js';
+import * as campaignRepository from '../../../prescription/campaign/infrastructure/repositories/campaign-repository.js';
+import { NoSkillsInCampaignError, NotFoundError } from '../../domain/errors.js';
+import { CampaignLearningContent } from '../../domain/models/CampaignLearningContent.js';
+import { LearningContent } from '../../domain/models/LearningContent.js';
+import * as areaRepository from './area-repository.js';
+import * as competenceRepository from './competence-repository.js';
+import * as frameworkRepository from './framework-repository.js';
+import * as skillRepository from './skill-repository.js';
+import * as thematicRepository from './thematic-repository.js';
+import * as tubeRepository from './tube-repository.js';
 
 async function findByCampaignId(campaignId, locale) {
   const skills = await campaignRepository.findSkills({ campaignId });
